Move new code to a dedicated file
